fix(mcp): give /mcps the same diagnostics as mcp list, and one CLI harness
`/mcps` reported only the per-server unresolved variables from discovery,
while `mcp list` also reported file-scoped blanks. A server templated as
`"url": "https://{env:MY_HOST}/mcp"` records against the config file
rather than the server, so with `MY_HOST` unset the CLI named it and the
session view said nothing — and the session view is where someone is
when a server will not connect. The wording is extracted into
`formatBlankedEnvForDisplay` so it is testable without standing up a
session; `/mcps` is otherwise only reachable through the whole handler.

The subprocess harness moves to `test/cli/fixtures/isolated-cli.ts`. It
was duplicated verbatim across the MCP CLI tests, and the duplication was
not cosmetic — each copy carried the `bun run --cwd` bug, so fixing one
left the other reading the repo's own config instead of the temp project.

That harness also swallowed spawn failures. `spawnSync` does not throw on
ENOENT or timeout; it returns `{ status: null, error }` with null stdout,
so a subprocess that never ran surfaced as `expected '' to contain
'broken'` and read like a test-logic bug. It now says the subprocess did
not complete, and why.

Full opencode suite: 11657 pass, 0 fail.
